fix and refactor configs: fix FrontendConfig with correct decorator, use RequiredArgsConstructor, remove front/back config interfaces, move to simplify folder structure
