+++ src/main/java/es/in2/vcverifier/component/CryptoComponent.java
@@ -4,7 +4,7 @@ import com.nimbusds.jose.crypto.bc.BouncyCastleProviderSingleton;
 import com.nimbusds.jose.jwk.Curve;
 import com.nimbusds.jose.jwk.ECKey;
 import com.nimbusds.jose.jwk.KeyUse;
-import es.in2.vcverifier.config.backend.BackendConfig;
+import es.in2.vcverifier.config.BackendConfig;
 import es.in2.vcverifier.exception.DidKeyCreationException;
 import es.in2.vcverifier.exception.ECKeyCreationException;
 import es.in2.vcverifier.util.UVarInt;
