--- src/main/java/es/in2/vcverifier/config/backend/BackendConfig.java
@@ -1,14 +1,0 @@
-package es.in2.vcverifier.config.backend;
-
-public interface BackendConfig {
-
-    String getUrl();
-
-    String getPrivateKey();
-
-    String getTrustedIssuerListUri();
-
-    String getClientsRepositoryUri();
-
-    String getRevocationListUri();
-}
