+++ src/main/java/es/in2/vcverifier/security/filters/CustomAuthorizationRequestConverter.java
@@ -4,8 +4,8 @@ import com.nimbusds.jose.Payload;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
 import es.in2.vcverifier.component.CryptoComponent;
+import es.in2.vcverifier.config.BackendConfig;
 import es.in2.vcverifier.config.CacheStore;
-import es.in2.vcverifier.config.backend.BackendConfig;
 import es.in2.vcverifier.model.AuthorizationContext;
 import es.in2.vcverifier.model.AuthorizationRequestJWT;
 import es.in2.vcverifier.service.DIDService;
