+++ src/test/java/es/in2/vcverifier/security/filters/CustomAuthenticationProviderTest.java
@@ -3,7 +3,6 @@ package es.in2.vcverifier.security.filters;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import es.in2.vcverifier.config.CacheStore;
-import es.in2.vcverifier.config.backend.BackendConfig;
 import es.in2.vcverifier.model.RefreshTokenDataCache;
 import es.in2.vcverifier.model.credentials.Issuer;
 import es.in2.vcverifier.model.credentials.lear.CredentialSubject;
