+++ src/test/java/es/in2/vcverifier/security/filters/CustomAuthorizationRequestConverterTest.java
@@ -5,7 +5,6 @@ import com.nimbusds.jose.jwk.ECKey;
 import com.nimbusds.jwt.SignedJWT;
 import es.in2.vcverifier.component.CryptoComponent;
 import es.in2.vcverifier.config.CacheStore;
-import es.in2.vcverifier.config.backend.BackendConfig;
 import es.in2.vcverifier.service.DIDService;
 import es.in2.vcverifier.service.JWTService;
 import jakarta.servlet.http.HttpServletRequest;
