+++ src/test/java/es/in2/vcverifier/service/ClientAssertionValidationServiceImplTest.java
@@ -2,7 +2,6 @@ package es.in2.vcverifier.service;
 
 import com.nimbusds.jose.Payload;
 import es.in2.vcverifier.config.JtiTokenCache;
-import es.in2.vcverifier.config.backend.BackendConfig;
 import es.in2.vcverifier.service.impl.ClientAssertionValidationServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
